requirements.txt
# Core
paddlepaddle==2.6.1 ; platform_system != "Darwin" or platform_machine != "arm64"
# If you're on Apple Silicon and want CPU-only, uncomment the next line and comment the line above:
# paddlepaddle==2.6.1 ; platform_system == "Darwin" and platform_machine == "arm64"
paddleocr==2.7.0.3

pdf2image==1.17.0
Pillow==10.4.0

# LangChain + Gemini
langchain==0.2.16
google-generativeai==0.8.3
langchain-google-genai==2.0.1

# Utils
python-dotenv==1.0.1
pydantic==2.9.2
rapidfuzz==3.9.7
dateparser==1.2.0
rich==13.9.2
tqdm==4.66.5c
typer==0.12.5
regex==2024.9.11
paddleocr
pdf2image
pillow
numpy
tqdm
python-dotenv
langchain
langchain-google-genai
typer
dateparser
pymupdf  

